+++ src/main/kotlin/com/fesi/flowit/common/auth/dto/TokenInfo.kt
@@ -1,0 +1,12 @@
+package com.fesi.flowit.common.auth.dto
+
+import java.util.*
+
+data class TokenInfo(
+    val email: String,
+    val userId: Long,
+    val issuedAt: Date,
+    val expiration: Date
+) {
+    companion object
+}
