+++ src/main/kotlin/com/fesi/flowit/common/auth/exception/InvalidUserException.kt
@@ -1,0 +1,3 @@
+package com.fesi.flowit.common.auth.exception
+
+class InvalidUserException : RuntimeException() {}
